A new look at state machine
