+++ backend/program/content/mdblist.py
@@ -1,9 +1,8 @@
 """Mdblist content module"""
-import json
 from utils.settings import settings_manager
 from utils.logger import logger
 from utils.request import RateLimitExceeded, RateLimiter, get, ping
-from program.media import MediaItemContainer
+from program.media.container import MediaItemContainer
 from program.updaters.trakt import Updater as Trakt
