+++ backend/program/scrapers/__init__.py
@@ -1,17 +1,11 @@
-import threading
 import time
-
 from utils.logger import logger
 from .torrentio import Torrentio
 
 
-class Scraping(threading.Thread):
-    def __init__(self, media_items):
-        super().__init__(name="Scraping")
-        self.media_items = media_items
-        self.services = [Torrentio(self.media_items)]
-        self.running = False
-        self.valid = False
+class Scraping():
+    def __init__(self):
+        self.services = [Torrentio()]
         while not self.validate():
             logger.error(
                 "You have no scraping services enabled, please enable at least one!"
@@ -21,18 +15,9 @@ class Scraping(threading.Thread):
     def validate(self):
         return any(service.initialized for service in self.services)
 
-    def run(self) -> None:
-        while self.running:
-            for service in self.services:
-                if service.initialized:
-                    service.run()
-                    time.sleep(1)
+    def run(self, item) -> None:
+        for service in self.services:
+            if service.initialized:
+                service.run(item)
 
-
-    def start(self) -> None:
-        self.running = True
-        super().start()
-
-    def stop(self) -> None:
-        self.running = False
-        super().join()
+scraper = Scraping()
