+++ backend/pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+minversion = 7.0
+; addopts = --cov=tests/.
+pythonpath = .
+testpaths =
+    tests
